+++ src/test/java/edu/hw1/task2/Task2Test.java
@@ -7,7 +7,7 @@ public class Task2Test {
 
     @Test
     void numberZero() {
-        int num = 0;
+        long num = 0;
         int actual = Task2.countDigits(num);
         int expected = 1;
         assertThat(actual).isEqualTo(expected);
@@ -15,7 +15,7 @@ public class Task2Test {
 
     @Test
     void oneDigit() {
-        int num = 5;
+        long num = 5;
         int actual = Task2.countDigits(num);
         int expected = 1;
         assertThat(actual).isEqualTo(expected);
@@ -23,7 +23,7 @@ public class Task2Test {
 
     @Test
     void twoDigits() {
-        int num = 34;
+        long num = 34;
         int actual = Task2.countDigits(num);
         int expected = 2;
         assertThat(actual).isEqualTo(expected);
@@ -31,7 +31,7 @@ public class Task2Test {
 
     @Test
     void threeDigits() {
-        int num = 544;
+        long num = 544;
         int actual = Task2.countDigits(num);
         int expected = 3;
         assertThat(actual).isEqualTo(expected);
@@ -39,7 +39,7 @@ public class Task2Test {
 
     @Test
     void fourDigits() {
-        int num = 4666;
+        long num = 4666;
         int actual = Task2.countDigits(num);
         int expected = 4;
         assertThat(actual).isEqualTo(expected);
@@ -47,7 +47,7 @@ public class Task2Test {
 
     @Test
     void negativeNumber() {
-        int num = -4666;
+        long num = -4666;
         int actual = Task2.countDigits(num);
         int expected = 4;
         assertThat(actual).isEqualTo(expected);
